cancel edit if it is the same
