+++ packages/cli/src/editor/components/RenderModal/SchemaEditor/ZodArrayEditor.tsx
@@ -13,6 +13,7 @@ import {Spacing} from '../../layout';
 import {ValidationMessage} from '../../NewComposition/ValidationMessage';
 import {optionRow} from '../layout';
 import {createZodValues} from './create-zod-values';
+import {deepEqual} from './deep-equal';
 import {useLocalState} from './local-state';
 import {SchemaFieldsetLabel} from './SchemaLabel';
 import type {JSONPath} from './zod-types';
@@ -155,34 +156,3 @@ export const ZodArrayEditor: React.FC<{
 		</div>
 	);
 };
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function deepEqual(a: any, b: any): boolean {
-	if (a === b) {
-		return true;
-	}
-
-	if (
-		typeof a !== 'object' ||
-		a === null ||
-		typeof b !== 'object' ||
-		b === null
-	) {
-		return false;
-	}
-
-	const keysA = Object.keys(a);
-	const keysB = Object.keys(b);
-
-	if (keysA.length !== keysB.length) {
-		return false;
-	}
-
-	for (const key of keysA) {
-		if (!keysB.includes(key) || !deepEqual(a[key], b[key])) {
-			return false;
-		}
-	}
-
-	return true;
-}
