+++ packages/cli/src/editor/components/RenderModal/SchemaEditor/local-state.ts
@@ -1,5 +1,6 @@
 import {useCallback, useRef, useState} from 'react';
 import type {z} from 'zod';
+import {deepEqual} from './deep-equal';
 import type {UpdaterFunction} from './ZodSwitch';
 
 type LocalState<T> = {
@@ -30,34 +31,28 @@ export const useLocalState = <T>({
 	const onChange = useCallback(
 		(updater: (oldV: T) => T, forceApply: boolean) => {
 			const newValue = updater(stateRef.current.value);
+			const isSame = deepEqual(newValue, stateRef.current.value);
+			if (isSame) {
+				return;
+			}
+
 			const safeParse = schema.safeParse(newValue);
 
 			if (safeParse.success || forceApply) {
 				setValue(updater, forceApply);
-				setLocalValue((oldLocalState) => {
-					const newState = {
-						revision: oldLocalState.revision + 1,
-						value: newValue,
-						zodValidation: safeParse,
-					};
-
-					stateRef.current = newState;
-
-					return newState;
-				});
-			} else {
-				setLocalValue((oldLocalState) => {
-					const newState = {
-						revision: oldLocalState.revision + 1,
-						value: newValue,
-						zodValidation: safeParse,
-					};
-
-					stateRef.current = newState;
-
-					return newState;
-				});
 			}
+
+			setLocalValue((oldLocalState) => {
+				const newState = {
+					revision: oldLocalState.revision + 1,
+					value: newValue,
+					zodValidation: safeParse,
+				};
+
+				stateRef.current = newState;
+
+				return newState;
+			});
 		},
 		[schema, setValue]
 	);
